Adds warehouse disk team113_wh01
Changes treemap so that it always starts from a '/' node

This could be adjusted so that rather than always creating a
root node, the merge code creates new higher level nodes only
when required when the merge is disjoint, such that the tree
would wind up with a root at the highest common path rather than
at '/'. Not implementing that for now.